# spring-quartz-scheduler-integration/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Customer - OpenAPI 3.0
  description: Everything about sample hibernate second level cache
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://petstore3.swagger.io/api/v1
tags:
  - name: email-data-store
    description: Everything about email
  - name: email-scheduler
    description: Everything about email-scheduler=
paths:
  /v1/emails/configs:
    post:
      tags:
        - email-data-store
      summary: Add an email data to database
      description: Add an email data to database
      operationId: saveEmailData
      requestBody:
        description: Add an email data to database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSchedulerDataRequest'
      responses:
        "201":
          description: successful created operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSchedulerDataResponse'
        "400":
          description: bad request operation
        "404":
          description: not found operation
        "500":
          description: internal server error operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/emails/configs/{email-config-id}:
    get:
      tags:
        - email-data-store
      summary: get an email config data to database
      description: get an email config data to database
      operationId: getEmailData
      parameters:
        - in: path
          name: email-config-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: successful created operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSchedulerDataResponse'
        "400":
          description: bad request operation
        "404":
          description: not found operation
        "500":
          description: internal server error operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - email-data-store
      summary: update an email data to database
      description: update an email data to database
      operationId: updateEmailData
      parameters:
        - in: path
          name: email-config-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Add an email data to database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSchedulerDataRequest'
      responses:
        "201":
          description: successful created operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSchedulerDataResponse'
        "400":
          description: bad request operation
        "404":
          description: not found operation
        "500":
          description: internal server error operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/emails/configs/{email-config-id}/actions/trigger-scheduler:
    post:
      tags:
        - email-scheduler
      summary: trigger email scheduler
      description: trigger email scheduler
      operationId: triggerEmailScheduler
      parameters:
        - in: path
          name: email-config-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: successful created operation
        "400":
          description: bad request operation
        "404":
          description: not found operation
        "500":
          description: internal server error operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    EmailSchedulerDataRequest:
      type: object
      properties:
        subject:
          type: string
        emailContent:
          type: string
        emailAddress:
          type: string
        daysOfWeek:
          type: array
          items:
            type: integer
            format: int32
        timesOfDay:
          type: array
          items:
            type: string
            format: time

    EmailSchedulerDataResponse:
      allOf:
        - $ref: '#/components/schemas/EmailSchedulerDataRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: datetime
            updatedAt:
              type: string
              format: datetime

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        status:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time